+++ test/remove_event_listener.test.ts
@@ -1,0 +1,20 @@
+import { removeEventListener, querySelector } from '../src';
+
+describe('removeEventListener', () => {
+  test('removeEventListener remove listener from element', () => {
+    document.body.innerHTML = '<p><p>';
+
+    let event: Event | null = null;
+
+    querySelector<HTMLParagraphElement>('p').map((p) => {
+      const listener = (e: Event) => (event = e);
+
+      p.addEventListener('click', listener);
+      removeEventListener(p, 'click', listener);
+
+      p.dispatchEvent(new Event('click'));
+
+      expect(event).toBeFalsy();
+    });
+  });
+});