+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Added
 
-- `addEventListener` function.
+- `addEventListener`, `removeEventListener` function.
 
 ## [0.2.0] - 2020-08-16